frontend/requirements.txt
streamlit
streamlit-option-menu
pillow
opencv-python
numpy
requests
datetime
folium
streamlit-folium